requirements.txt
tensorflow-gpu==1.14
numpy
matplotlib
tqdm
yaml
